Generalise quantification algorithm
