+++ src/adiar/CMakeLists.txt
@@ -42,6 +42,7 @@ set(HEADERS
   internal/pred.h
   internal/levelized_priority_queue.h
   internal/product_construction.h
+  internal/quantify.h
   internal/reduce.h
 )
